+++ clang/lib/Serialization/ASTReader.cpp
@@ -11508,6 +11508,9 @@ OMPClause *OMPClauseReader::readClause() {
   case llvm::omp::OMPC_partial:
     C = OMPPartialClause::CreateEmpty(Context);
     break;
+  case llvm::omp::OMPC_depth:
+    C = new (Context) OMPDepthClause();
+    break;
   case llvm::omp::OMPC_looprange:
     C = OMPLoopRangeClause::CreateEmpty(Context);
     break;
@@ -11946,6 +11949,11 @@ void OMPClauseReader::VisitOMPFullClause(OMPFullClause *C) {}
 
 void OMPClauseReader::VisitOMPPartialClause(OMPPartialClause *C) {
   C->setFactor(Record.readSubExpr());
+  C->setLParenLoc(Record.readSourceLocation());
+}
+
+void OMPClauseReader::VisitOMPDepthClause(OMPDepthClause *C) {
+  C->setDepth(Record.readSubExpr());
   C->setLParenLoc(Record.readSourceLocation());
 }